% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briteParser.R
\name{.briteParser}
\alias{.briteParser}
\title{Parses a a list of BRITE entries}
\usage{
.briteParser(briteCol)
}
\arguments{
\item{briteCol}{List with BRITE entries}
}
\description{
Parses a a list of BRITE entries
}
\keyword{internal}
