% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChemCompounds.R
\name{.extractHTML}
\alias{.extractHTML}
\title{Extracts compounds from PubChem, and KEGG when unavailable in the earlier}
\usage{
.extractHTML(htmlFile, node1, node2)
}
\arguments{
\item{htmlFile}{Data extracted from read_html}

\item{node1}{First node to search on the XML}

\item{node2}{Second node to search on the XML}
}
\description{
Extracts compounds from PubChem, and KEGG when unavailable in the earlier
}
