% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChemCompounds.R
\name{.getCIDsProperties}
\alias{.getCIDsProperties}
\title{Extracts compounds properties from CIDs}
\usage{
.getCIDsProperties(CIDs, pubChemIDswCID)
}
\arguments{
\item{CIDs}{Character vector with valid CIDs}
}
\description{
Extracts compounds properties from CIDs
}
