% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{.getCompoundPathways}
\alias{.getCompoundPathways}
\title{Extracts compounds data from KEGG}
\usage{
.getCompoundPathways(pathwayList)
}
\arguments{
\item{pathwayList}{List containing data concerning the pathways}
}
\description{
Extracts compounds data from KEGG
}
