% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{.getCompoundsPathways}
\alias{.getCompoundsPathways}
\title{Extracts the pathways that contain information regarding the compounds participating on it}
\usage{
.getCompoundsPathways(pathwayInfo)
}
\arguments{
\item{pathwayInfo}{List containing data concerning the pathways}
}
\description{
Extracts the pathways that contain information regarding the compounds participating on it
}
