% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{.getKeggCompounds}
\alias{.getKeggCompounds}
\title{Extracts compounds data from KEGG}
\usage{
.getKeggCompounds(compounds)
}
\arguments{
\item{compounds}{Character vector with valid KEGG IDs}
}
\description{
Extracts compounds data from KEGG
}
