% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{.getKeggProperties}
\alias{.getKeggProperties}
\title{Extracts compounds properties from KEGG}
\usage{
.getKeggProperties(listOfKegg, property)
}
\arguments{
\item{listOfKegg}{List with data from compounds, extracted from KEGG.}

\item{property}{List of properties to extract.}
}
\description{
Extracts compounds properties from KEGG
}
