% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKnapSackCompounds.R
\name{.getKnapSackCompounds}
\alias{.getKnapSackCompounds}
\title{Extracts compounds from KNApSAcK}
\usage{
.getKnapSackCompounds(item, keyword, species)
}
\arguments{
\item{item}{Informs the type of information to be provided to the website.
\itemize{
\item \code{Organism};
\item \code{Metabolite};
\item \verb{Molecular formula};
\item ...
}}

\item{keyword}{Specifies what is to be searched}

\item{species}{Species' name}
}
\description{
Extracts compounds from KNApSAcK
}
