% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{.getPathwayInfo}
\alias{.getPathwayInfo}
\title{Get informations regaring pathways of the chosen organism}
\usage{
.getPathwayInfo(pathwaysID)
}
\arguments{
\item{pathwaysID}{Character vector with valid KEGG IDs for pathways.}
}
\description{
Get informations regaring pathways of the chosen organism
}
