% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPubChemCompounds.R
\name{.getPubChemCompounds}
\alias{.getPubChemCompounds}
\title{Extracts compounds from PubChem, and KEGG when unavailable in the earlier}
\usage{
.getPubChemCompounds(compounds)
}
\arguments{
\item{compounds}{Character vector with valid KEGG IDs}
}
\description{
Extracts compounds from PubChem, and KEGG when unavailable in the earlier
}
