% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R
\name{getOrganismName}
\alias{getOrganismName}
\title{-- FOR FUTURE USE ON SHINY APP -- Displays the organism name corresponding to a given KEGG code}
\usage{
getOrganismName(orgName)
}
\arguments{
\item{orgName}{KEGG code of the organism of interest.}
}
\value{
A single string with the organism name.
}
\description{
Given a KEGG code, displays the name of an organism
}
\examples{
#Retrieve available organisms from KEGG
df <- whichOrganisms(x = "organism")

#Get name from the KEGG ID for Vitis vinifera
getOrganismCode(orgName = "vvi")
}
