% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R
\name{getOrganismNamesCode}
\alias{getOrganismNamesCode}
\title{Get all organisms' names available on KEGG}
\usage{
getOrganismNamesCode()
}
\value{
A data.frame vector containing all organisms' names available in KEGG.
}
\description{
Simple function to provide all available organisms present in the KEGG database.
}
