% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R
\name{whichOrganisms}
\alias{whichOrganisms}
\title{Retrieves organisms available on KEGG}
\usage{
whichOrganisms(x = "organism")
}
\arguments{
\item{x}{Parameter to specify what KEGGREST should look on KEGG.}
}
\value{
A \code{data.frame} object.
}
\description{
Creates a \code{data.frame} with the KEGG identifiers and scientific names of all organisms
available in the database.
}
\examples{
df <- whichOrganisms(x = "organism")
}
