% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompoundTable.R
\name{getCompoundTable}
\alias{getCompoundTable}
\title{Creates the compound database for a specific organism, from PubChem or KEGG, as specified by the user}
\usage{
getCompoundTable(orgCode, db)
}
\arguments{
\item{orgCode}{KEGG ID for the organism of interest. See \code{\link[=whichOrganisms]{whichOrganisms()}} for a list of available organisms.}

\item{db}{Database from where the compounds should be extracted.
\itemize{
\item \code{db = "KEGG"}: Extract compounds from \href{https://www.kegg.jp}{KEGG} database.
\item \code{"db = PubChem"}: Extract compounds from \href{https://pubchem.ncbi.nlm.nih.gov}{PubChem} database.
}}
}
\value{
A \code{data.frame} containing, for each compound:
\itemize{
\item the KEGG identifier;
\item the PubChem identifier (when \code{db = "PubChem"});
\item its name;
\item molecular formula;
\item exact mass;
\item and the pathways it participates in.
}
}
\description{
\code{getCompoundTable()} creates a dataframe object containing information regarding the compounds
participating in a user specified organism, as long as this information is available in Kyoto
Encyclopedia of Genes and Genomes data (KEGG) database.

Although it primarily relies on KEGG, the function can also retrieve compound information from the PubChem
database provided that these compounds are cross referenced. Using PubChem has the advantage
of being faster and, in some cases, resulting in more standardised molecular formulas.
}
\details{
Compounds with no available information in regards to their name, molecular formula and/or exact
mass are promptly removed from the final \code{data.frame}.
}
\examples{
## Get compounds availabe for Vitis vinifera from PubChem
vviData <- getCompoundTable("vvi", "PubChem")
}
