% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableData.R
\name{getOrganismName}
\alias{getOrganismName}
\title{Get all organisms' names available on KEGG}
\usage{
getOrganismName()
}
\value{
A chr vector containing all organisms' names
}
\description{
Simple function to provide all available organisms present in the KEGG database.
}
